Added new tests for close case actions on closed cases
Added new tests for close case actions on closed cases
