+++ ConcernsCaseWork/ConcernsCaseWork.CypressTests/cypress/integration/casework-regression/closed-cases.js
@@ -1,3 +1,11 @@
+import AddToCasePage from "/cypress/pages/caseActions/addToCasePage";
+import CaseManagementPage from "/cypress/pages/caseMangementPage";
+import utils from "/cypress/support/utils"
+import ntiAddPage from "/cypress/pages/caseActions/ntiAddPage";
+import CaseActionsBasePage from "/cypress/pages/caseActions/caseActionsBasePage";
+import homePage from "/cypress/pages/homePage.js";
+import closedCasePage from "/cypress/pages/closedCasePage.js";
+
 describe('User can view and navigate Closed cases', () => {
     before(() => {
 		cy.login();
@@ -7,27 +15,26 @@ describe('User can view and navigate Closed cases', () => {
 		cy.storeSessionData();
 	});
 
-    it('User clicks on Create Case and should see Search Trusts', () => {
-        cy.get('[href="/case/closed"]').click()
-        cy.get('.govuk-table__body tr').should('be.visible')
+    it('User clicks Closed Case button and is taken to closed case page', () => {    
+        homePage.getClosedCasesBtn().click();
+        closedCasePage.getClosedCasesTable().should('be.visible');
     });
 
     it('User can view a closed case', () => {
-        cy.get('[href*=closed]').first().should('be.visible')
-        cy.get('[href*=closed]').first().click()
-        cy.get('[name=caseID]').should('be.visible')
+        closedCasePage.getClosedCasesLink().first().should('be.visible');
+        closedCasePage.getClosedCasesLink().first().click();
+        CaseManagementPage.getCaseID().should('be.visible');
     });
 
-    it('User clicks Back  and should be taken back to the Closed Cases', () => {
-        cy.get('#back-link-event').click()
-        cy.get('.govuk-table__body tr').should('be.visible')
+    it('User clicks Back and should be taken back to the Closed Cases', () => {
+        CaseManagementPage.getBackBtn().click();
+        closedCasePage.getClosedCasesTable().should('be.visible');
     });
     
-    it('User clicks Back  and should be taken back to the Active Casework screen', () => {
-        cy.get('#back-link-event').click()
-        cy.get('[href="/case"').should('be.visible')
-        cy.get('[href="/trust"').should('be.visible')
-        cy.get('[href="/case/closed"').should('be.visible')
+    it('User clicks Back and should be taken back to the Active Casework screen', () => {
+        CaseManagementPage.getBackBtn().click();
+        homePage.getHeadingText().should('be.visible');
+        homePage.getClosedCasesBtn().should('be.visible');
     });
